Main check-in for #18. Wrapped most methods and variables as properties of the new gSALR object. Updated references to these.
Menu and quick quote stuff is still outside of gSALR for now since they both need some overhauling.

Removed xpc unwrapping from the overlay as it seems that Firefox is doing all that stuff consistently now. Might need to drop 3.0.X compatibility as a result, but that should be fine since it was EOL early this year. Without the extra unwrapping, SALR will be more secure.

Added support for the new forum method of displaying mod/admin stars in threads.

Added support for opening a changelog tab after upgrading SALR.